requirements.txt
Flask==3.0.0
PyJWT==2.9.0
passlib==1.7.4
python-dotenv==1.0.1